requirements.txt
aiogram==3.22.0
python-dotenv==1.2.1
pydantic==2.11.10
fluent-runtime==0.4.0
pydantic-settings==2.11.0
asyncpg==0.30.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11